Move parser,resolution error classes to parsing,resolution dirs
---
